+++ R/distance2_methods.R
@@ -183,8 +183,9 @@ distance2randomforest <- function(formula, data = NULL, link = NULL, ...) {
   res <- randomForest::randomForest(formula, data = newdata, ...)
   return(list(model = res, distance = predict(res, type = "prob")[,"1"]))
 }
-#distance2lasso--------------
-distance2glmnet <- function(formula, data = NULL, link = NULL, ...) {
+
+#distance2glmnet--------------
+distance2elasticnet <- function(formula, data = NULL, link = NULL, ...) {
   if (!is.null(link) && startsWith(as.character(link), "linear")) {
     linear <- TRUE
     link <- sub("linear.", "", as.character(link), fixed = TRUE)
@@ -200,6 +201,8 @@ distance2glmnet <- function(formula, data = NULL, link = NULL, ...) {
   else if (link == "log") A$family <- "poisson"
   else A$family <- binomial(link = link)
 
+  if (is.null(A[["alpha"]])) A[["alpha"]] <- .5
+
   mf <- model.frame(formula, data = data)
 
   A$y <- model.response(mf)
@@ -214,6 +217,71 @@ distance2glmnet <- function(formula, data = NULL, link = NULL, ...) {
 
   return(list(model = res, distance = pred))
 }
+distance2lasso <- function(formula, data = NULL, link = NULL, ...) {
+  if (!is.null(link) && startsWith(as.character(link), "linear")) {
+    linear <- TRUE
+    link <- sub("linear.", "", as.character(link), fixed = TRUE)
+  }
+  else linear <- FALSE
+
+  A <- list(...)
+  s <- A[["s"]]
+  A[!names(A) %in% c(names(formals(glmnet::glmnet)), names(formals(glmnet::cv.glmnet)))] <- NULL
+
+  if (is.null(link)) link <- "logit"
+  if (link == "logit") A$family <- "binomial"
+  else if (link == "log") A$family <- "poisson"
+  else A$family <- binomial(link = link)
+
+  A[["alpha"]] <- 1
+
+  mf <- model.frame(formula, data = data)
+
+  A$y <- model.response(mf)
+  A$x <- model.matrix(update(formula, . ~ . + 1), mf)[,-1,drop = FALSE]
+
+  res <- do.call(glmnet::cv.glmnet, A)
+
+  if (is.null(s)) s <- "lambda.1se"
+
+  pred <- drop(predict(res, newx = A$x, s = s,
+                       type = if (linear) "link" else "response"))
+
+  return(list(model = res, distance = pred))
+}
+distance2ridge <- function(formula, data = NULL, link = NULL, ...) {
+  if (!is.null(link) && startsWith(as.character(link), "linear")) {
+    linear <- TRUE
+    link <- sub("linear.", "", as.character(link), fixed = TRUE)
+  }
+  else linear <- FALSE
+
+  A <- list(...)
+  s <- A[["s"]]
+  A[!names(A) %in% c(names(formals(glmnet::glmnet)), names(formals(glmnet::cv.glmnet)))] <- NULL
+
+  if (is.null(link)) link <- "logit"
+  if (link == "logit") A$family <- "binomial"
+  else if (link == "log") A$family <- "poisson"
+  else A$family <- binomial(link = link)
+
+  A[["alpha"]] <- 0
+
+  mf <- model.frame(formula, data = data)
+
+  A$y <- model.response(mf)
+  A$x <- model.matrix(update(formula, . ~ . + 1), mf)[,-1,drop = FALSE]
+
+  res <- do.call(glmnet::cv.glmnet, A)
+
+  if (is.null(s)) s <- "lambda.1se"
+
+  pred <- drop(predict(res, newx = A$x, s = s,
+                       type = if (linear) "link" else "response"))
+
+  return(list(model = res, distance = pred))
+}
+
 #distance2gbm--------------
 distance2gbm <- function(formula, data = NULL, link = NULL, ...) {
   if (!is.null(link) && startsWith(as.character(link), "linear")) {